fix: fix linters errors
