+++ student.rb
@@ -6,7 +6,7 @@ class Student < Person
   def initialize(age, parent_permision = 'true', name = 'unknown', classroom = nil)
     @classroom = classroom
     unless classroom.nil?
-      (classroom.students.push(self) unless classroom.students.include?(self)) 
+      (classroom.students.push(self) unless classroom.students.include?(self))
     end
     super(age, name, parent_permision)
   end
